requirements.txt
pydantic>=2.6,<3
typer>=0.12,<1
rich>=13.7,<14
google-generativeai>=0.8.0
openai>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
requests
python-pptx
duckduckgo-search>=5.0
